improve style and remove possible bug

+++ 03_Blatt/utils.cpp
@@ -1,7 +1,3 @@
-#include "utils.hpp"
-#include <iostream>
-
-
 bool isPermutation(int* perm, int count)
 {
     // for 1-n , dann lineare suche
@@ -24,8 +20,8 @@ bool isSorted(double* data, int dataCount, int* perm)
 {
     // for 1-n , dann lineare suche
     for (int i = 1; i < dataCount; ++i) {
-        int lesser = 0;
-        int bigger = 0;
+        double lesser = 0;
+        double bigger = 0;
         for (int j = 0; j < dataCount; ++j) {
             if (perm[j] == i) {
                 bigger = data[j];
@@ -34,7 +30,7 @@ bool isSorted(double* data, int dataCount, int* perm)
                 lesser = data[j];
             }
         }
-        if (lesser>bigger) {
+        if (lesser > bigger) {
             return false;
         }
     }
